Refactor: Binary Tree algorithm
The BinaryTreeMazeGenerator's template was removed since the width and
height can be obtained from the provided grid.

Resolves: #PMG-3
